global rewrite types int -> uint32_t

+++ kfft_bfly.c
@@ -1,7 +1,7 @@
 static inline void
-kf_bfly2(kfft_cpx* Fout, const size_t fstride, const kfft_kplan_t* st, int m) {
+kf_bfly2(kfft_cpx* Fout, const uint32_t fstride, const kfft_kplan_t* st, uint32_t m) {
     kfft_cpx* Fout2;
-    size_t twidx = 0;
+    uint32_t twidx = 0;
     kfft_cpx t;
     Fout2 = Fout + m;
     do {
@@ -15,12 +15,12 @@ kf_bfly2(kfft_cpx* Fout, const size_t fstride, const kfft_kplan_t* st, int m) {
 }
 
 static inline void
-kf_bfly4(kfft_cpx* Fout, const size_t fstride, const kfft_kplan_t* st, const size_t m) {
-    size_t tw1, tw2, tw3;
+kf_bfly4(kfft_cpx* Fout, const uint32_t fstride, const kfft_kplan_t* st, const uint32_t m) {
+    uint32_t tw1, tw2, tw3;
     kfft_cpx scratch[6];
-    size_t k = m;
-    const size_t m2 = 2 * m;
-    const size_t m3 = 3 * m;
+    uint32_t k = m;
+    const uint32_t m2 = 2 * m;
+    const uint32_t m3 = 3 * m;
 
     tw1 = tw2 = tw3 = 0;
 
@@ -55,10 +55,10 @@ kf_bfly4(kfft_cpx* Fout, const size_t fstride, const kfft_kplan_t* st, const siz
 }
 
 static inline void
-kf_bfly3(kfft_cpx* Fout, const size_t fstride, const kfft_kplan_t* st, size_t m) {
-    size_t k = m;
-    const size_t m2 = 2 * m;
-    size_t tw1, tw2;
+kf_bfly3(kfft_cpx* Fout, const uint32_t fstride, const kfft_kplan_t* st, uint32_t m) {
+    uint32_t k = m;
+    const uint32_t m2 = 2 * m;
+    uint32_t tw1, tw2;
     kfft_cpx scratch[5];
     kfft_cpx epi3;
     epi3 = TWIDDLE(fstride * m, st);
@@ -92,9 +92,9 @@ kf_bfly3(kfft_cpx* Fout, const size_t fstride, const kfft_kplan_t* st, size_t m)
 }
 
 static inline void
-kf_bfly5(kfft_cpx* Fout, const size_t fstride, const kfft_kplan_t* st, int m) {
+kf_bfly5(kfft_cpx* Fout, const uint32_t fstride, const kfft_kplan_t* st, uint32_t m) {
     kfft_cpx *Fout0, *Fout1, *Fout2, *Fout3, *Fout4;
-    int u;
+    uint32_t u;
     kfft_cpx scratch[13];
     kfft_cpx ya, yb;
     ya = TWIDDLE(fstride * m, st);
